Drop support for Rails 4.2
Attributes API has been changed significantly in Rails 5.0, and this
implementation is not compatible with Rails 4.2.  Personally, I don't
see much value in fixing it, so I simply remove it.

People are suggested to upgrade to Rails 5, or to use the original
gem.  Nevertheless, I've opened an issue in case someone wants to
contribute (https://github.com/riboseinc/activeuuid/issues/28).