Constrain SQLite3 gem version
Active Record has version constraint on SQLite3 gem.  Unfortunately, it
is not specified in gemspec, but in a source file where adapter class is
defined, therefore Bundler is unaware of it.  Adding an explicit version
constraint in this gem's gemspec guarantees that only compatible version
is installed.

It is expected that dependency version constraint will be relaxed in
Active Record 5.2.3.

See this SO answer for more details:
https://stackoverflow.com/a/54729071/304175.